pyproject.toml
[project]
name = "api-semantic"
version = "0.1.0"
description = "FastAPI service for text embeddings"
readme = "README.md"
authors = [{ name = "Christian Fita", email = "you@example.com" }]
requires-python = ">=3.10,<3.13"
dependencies = [
  "fastapi>=0.116.0,<0.117.0",
  "uvicorn[standard]>=0.35.0,<0.36.0",
  "sentence-transformers>=3.4.1,<4.0.0",
  "numpy>=2.2.3,<3.0.0",
  "pydantic-settings>=2.8.1,<3.0.0",
]

[tool.poetry]
packages = [{ include = "api_semantic", from = "src" }]

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
httpx = "^0.28.1"

[build-system]
requires = ["poetry-core>=1.8.0"]
build-backend = "poetry.core.masonry.api"
